requirements.txt
Pillow==9.2.0
PyMySQL==1.0.2
